Add -Quiet parameter to New-TeamCityBlock
The presence of this switch modifies the behavior of the Cmdlet such
that it will not use Write-Host and Write-Output as separate output
streams.  When set, the message will be written using Write-Output and
nothing will be written using Write-Host.  This change was necessary
solely for the sake of testing.